Implement getKeyStats
Note that there are some values that return null, which cannot be
represented in C++ primitive types such as int and double. In this case
we arbitrarily choose zero, which admittedly is not the same as null and
may in fact be a valid (or invalid) numeric value in context.

Addresses #3.